+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add single-day planning column to Oggi view
+
 ## [0.7.0] - 2026-07-06
 
 ### Added